requirements.txt
torch
torchvision
segmentation-models-pytorch
opencv-python
rasterio
shapely
geopandas
pyproj
albumentations
numpy
pandas
pillow
requests
pyyaml
scipy
wandb
fastapi
uvicorn
tqdm
matplotlib
